Moved functions relating to aligning and offsets between cubes into seperate .py file
